requirements.txt
django
Pillow
redis
djangorestframework
requests
channels
channels-redis
psycopg2-binary
uwsgi
daphne